feat(product): add product

+++ Entities/Orders.cs
@@ -8,7 +8,7 @@ namespace Shop_BE.Entities
         public int OrderId { get; set; }
         public int UserId { get; set; }
         public string OrderDate { get; set; }
-        public float TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
         public string Status { get; set; }
         public string Address { get; set; }
