+++ Entities/Products.cs
@@ -8,8 +8,8 @@ namespace Shop_BE.Entities
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
         public string? Description { get; set; }
-        public float? Price { get; set; }
-        public float? OldPrice { get; set; }
+        public decimal? Price { get; set; }
+        public decimal? OldPrice { get; set; }
         public int? CategoryId { get; set; }
         public int? PromotionId { get; set; }
     }
