+++ Entities/Promotions.cs
@@ -7,7 +7,7 @@ namespace Shop_BE.Entities
         [Key]
         public int? PromotionId { get; set; }
         public string? PromotionName { get; set; }
-        public float? Discount { get; set; }
+        public decimal? Discount { get; set; }
         public string? StartDate { get; set; }
         public string? EndDate { get; set; }
     }
