+++ Request/AddProduct.cs
@@ -4,8 +4,8 @@
     {
         public string ProductName { get; set; }
         public string Description { get; set; }
-        public float Price { get; set; }
-        public float OldPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal OldPrice { get; set; }
         public int CategoryId { get; set; }
         public List<IFormFile> Images { get; set; }
